feat(sandbox): extract copyEnemyJsonStubToClipboard to sandbox_export.js and wire it up in sandbox_panel.js
